+++ src/variable.c
@@ -544,8 +544,7 @@ int translate_color_names(struct session *ses, char *string, char *result)
 
 	while (*string)
 	{
-
-		if (isalpha(*string))
+		if (isalpha((int) *string))
 		{
 			for (cnt = 0 ; *color_table[cnt].name ; cnt++)
 			{
@@ -619,7 +618,7 @@ int get_color_names(struct session *ses, char *string, char *result)
 
 	while (*string)
 	{
-		if (isalpha(*string))
+		if (isalpha((int) *string))
 		{
 			for (cnt = 0 ; *color_table[cnt].name ; cnt++)
 			{
